feat: pegando dados API

+++ Wishlist.API/Wishlist.API/DTO/ProductDTO.cs
@@ -4,12 +4,13 @@ namespace Wishlist.API.DTO
     [Serializable]
     public class ProductDTO
     {
-        public decimal Price { get; set; }
+        public float Price { get; set; }
         public string Image { get; set; }
         public Guid Id { get; set; }
         public string Brand { get; set; }
         public string Title { get; set; }
-        public decimal? ReviewScore { get; set; }
+        public float? ReviewScore { get; set; }
 
     }
+
 }
